Disable sampling for nlocations=0
In case the user does not know how many mutation locations are existing
and they want to run mutation for all of them, using a value of `0` for
the `nlocations` parameter seems like a natural choice, especially since
many tools treat a value of `0` for a limit as infinity/all possible.

Also, the exception message in the `get_mutation_sample_locations`
function suggests that `0` might be a reasonable value.  Obviously,
taking an empty sample with this value does not make too much sense, as
this would disable mutation at all.